Refactor before implementing `retry` + minor fixes
I started implementing `retry`, but realized my general strategy was
wrong. I was trying to retry "in place" using the current job, but that
made it hard to keep the database and the corresponding Kubernetes jobs
in sync, because there might be multiple Kubernetes jobs for each
falconeri job.

But this portion of the code can still be salvaged from my first
attempt.

For my next attempt, I'm going to use the existing falconeri job to
create a new job that contains just the necessary datums.