Rename `prefix` module to `prelude` (per Rust convention)
A `prelude` module in Rust is a module which contains a set of helpful
imports that are normally needed to use an API, and which is normally
wildcard imported as `use some_crate::prelude::*;`. A prelude does not
include rarely-used names.

The `falconeri_common` crate used a non-standard name for this for some
reason.