falconeri-worker: Port to use REST API
All worker communication should now go via HTTP to `falconerid` instead
of directly to `falconeri-postgres`.

This should make it much easier to scale from ~175 to 500+ CPUs, because
it will sharply reduce the number of database connections we try to
open.

Of course, we'll end up managing huge numbers of HTTP connections
instead, but at least there are tools for that. (And we can probably
just turn off HTTP keepalive and _maybe_ add a proxy server in front of
`falconerid` if absolutely necessary. It's all in-cluster so new
connections should be fairly cheap.)