Deploy `falconerid` alongside PostgreSQL
We now want to have two copies of `falconerid` running on our cluster,
so that the new `falconeri-worker` can talk to them.

Part of the challenge of setting this up was making sure we had some way
to test it using a local `minikube` cluster so that we could iterate on
the various pieces more rapidly. This is done using a `falconeri deploy
--development` flag, which customizes the YAML manifest we pass to
`kubectl apply -f`.