v0.2.0: Increase scalability using HTTP
This version of falconeri should significantly improve scalability by
replacing `falconeri-worker`'s connections to the database with HTTP
connections to `falconcerid`. We've tested this with 300 workers with no
problems, and we anticipate that it should be able to reach 500
workers (perhaps with some slight tuning).

NOTE: To update from an earlier version of falconeri, you must:

1. Run `falconeri undeploy` using the old version.
2. Install the new version of `falconeri`.
3. Run `falconeri deploy` using the new version.

This will preserve your existing database and credentials. The
undeployment is required because we replace the `falconeri-postgres`
replica set with a full-fledged deployment.