Untested: Create HTTP API for most important client operations
The unit tests pass but the integration tests don't yet.

This is a big refactoring, moving a lot of code into `falconerid` and
adding HTTP client and server layers to access it. In particular, this
removes all database access from the client for `job run`, `run retry`,
and the new `job_wait`.

Things left to do:

- Test this code.
- Make sure that `falconerid` is installed with a Kubernetes role that
  allows it to create job resources on the cluster.
- Decide if `job wait` should output JSON, or if something else should.