retry: Explain why we need a second client for now
This is a limitation in our `Client` type and in how we handle `retry`.
We should fix `retry` to be truly idempotent on the server (by having
the client send a UUID with each request that we can use to detect
duplicate requests), and then wrap the client in a retry block.