Implement basic babysitter & retries
We add support for `datum_tries`. This involves adding a "babysitter"
thread to each running copy of `falconerid`. The babysitter wakes up
every two minutes, and looks at the running datums and jobs. If
anything looks broken, the babysitter fixes it.

Importantly, the babysitter is stateless and it relies on PostgreSQL
locks to coordinate all operations. So it's normal to run multiple
copies of the babysitter. And it should remain possible (and safe) to
restart `falconerid` at any time.